requirements.txt
toga-winforms
requests
aiohttp
qrcode
pyperclip